# openapi.yaml
openapi: 3.1.0
info:
  title: Insight Store — Telegram Forwarding Bot
  description: Automatic Telegram forwarding between authorized chats. No quotas,
    no subscription limits — and no evasion of Telegram's own limits.
  version: 0.1.0
paths:
...
  /api/v1/activity:
    get:
      tags:
      - status
      summary: Activity
      operationId: activity_api_v1_activity_get
      parameters:
      - name: outcome
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Outcome
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EventResponse'
                title: Response Activity Api V1 Activity Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/usage/summary:
    get:
      tags:
      - status
      summary: Usage Summary
      description: Operational counters only. This product has no quotas to account
        for.
      operationId: usage_summary_api_v1_usage_summary_get
      parameters:
      - name: period
        in: query
        required: false
        schema:
          type: string
          pattern: ^(24h|7d|30d)$
          default: 24h
          title: Period
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsageSummaryResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/audit-events:
    get:
      tags:
      - status
      summary: Audit Events
      operationId: audit_events_api_v1_audit_events_get
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AuditEventResponse'
                title: Response Audit Events Api V1 Audit Events Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
...
    AuditEventResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        action:
          type: string
          title: Action
        object_type:
          type: string
          title: Object Type
        object_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Object Id
        correlation_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Correlation Id
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - action
      - object_type
      - object_id
      - correlation_id
      - created_at
      title: AuditEventResponse
...
    EventResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        rule_id:
          type: string
          format: uuid
          title: Rule Id
        outcome:
          type: string
          title: Outcome
        reason_code:
          type: string
          title: Reason Code
        reason_text:
          type: string
          title: Reason Text
          default: ''
        detail_safe:
          anyOf:
          - type: string
          - type: 'null'
          title: Detail Safe
        attempt:
          type: integer
          title: Attempt
        source_chat_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Source Chat Id
        destination_chat_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Destination Chat Id
        source_message_ids:
          items:
            type: string
          type: array
          title: Source Message Ids
        occurred_at:
          type: string
          format: date-time
          title: Occurred At
      type: object
      required:
      - id
      - rule_id
      - outcome
      - reason_code
      - detail_safe
      - attempt
      - source_chat_id
      - destination_chat_id
      - occurred_at
      title: EventResponse
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    UsageSummaryResponse:
      properties:
        period:
          type: string
          title: Period
        forwarded:
          type: integer
          title: Forwarded
          default: 0
        skipped:
          type: integer
          title: Skipped
          default: 0
        failed:
          type: integer
          title: Failed
          default: 0
        retry_scheduled:
          type: integer
          title: Retry Scheduled
          default: 0
        paused:
          type: integer
          title: Paused
          default: 0
      type: object
      required:
      - period
      title: UsageSummaryResponse
      description: 'Operational counters. Explicitly not quota accounting — this product
        has

        no quotas.'
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...